+++ src/cpp/impl/CellSim.Cells.CellType.cpp
@@ -2,27 +2,39 @@
 
 namespace CellSim::Cells
 {
-    ::std::vector<::std::string> CellType::s_names{ "Invalid" };
+    ::std::vector<::std::pair<::std::string, Graphics::Color>> CellType::s_names{ {"Invalid", { 0, 0, 0, 0 }} };
 
-    CellType CellType::AddName(::std::string name)
+    CellType CellType::AddName(::std::string name, Graphics::Color color)
     {
         // すでに名前が存在すればそのIdを返す
         for (size_t index = 0;index != s_names.size(); ++index) {
-            if (s_names[index] == name) [[unlikely]] return CellType((uint32_t)index);
+            if (s_names[index].first == name) [[unlikely]] return CellType((uint32_t)index);
         }
 
-        s_names.emplace_back(::std::move(name));
+        s_names.emplace_back(::std::move(name), color);
 
         return CellType((uint32_t)s_names.size());
     }
 
-    CellType CellType::AddName(::std::string_view name)
+    CellType CellType::AddName(::std::string_view name, Graphics::Color color)
     {
         // すでに名前が存在すればそのIdを返す
         for (size_t index = 0;index != s_names.size(); ++index) {
-            if (s_names[index] == name) [[unlikely]] return CellType((uint32_t)index);
+            if (s_names[index].first == name) [[unlikely]] return CellType((uint32_t)index);
         }
 
-        s_names.emplace_back(name);
+        s_names.emplace_back(name, color);
+
+        return CellType((uint32_t)s_names.size());
+    }
+
+    CellType CellType::FromName(::std::string_view name) noexcept
+    {
+        // すでに名前が存在すればそのIdを返す
+        for (size_t index = 0;index != s_names.size(); ++index) {
+            if (s_names[index].first == name) [[unlikely]] return CellType((uint32_t)index);
+        }
+
+        return Invalid();
     }
 }