+++ src/cpp/impl/CellSim.Cells.CellType.hpp
@@ -2,6 +2,7 @@
 #define CELLSIM_CELLS_CELLTYPE_HPP
 
 #include "base.hpp"
+#include "CellSim.Graphics.Color.hpp"
 #include <string>
 #include <string_view>
 #include <vector>
@@ -13,7 +14,7 @@ namespace CellSim::Cells
     class CellType final {
         private:
 
-        static ::std::vector<::std::string> s_names;
+        static ::std::vector<::std::pair<::std::string, Graphics::Color>> s_names;
 
         uint32_t m_id;
 
@@ -26,19 +27,27 @@ namespace CellSim::Cells
         /// @brief 種類を追加
         /// @param name 名前
         /// @return その名前に対応するCellType値
-        [[nodiscard]] static CellType AddName(const char* name);
+        [[nodiscard]] static CellType AddName(const char* name, Graphics::Color color);
 
         /// @brief 種類を追加
         /// @param name 名前
         /// @return その名前に対応するCellType値
-        [[nodiscard]] static CellType AddName(::std::string name);
+        [[nodiscard]] static CellType AddName(::std::string name, Graphics::Color color);
 
         /// @brief 種類を追加
         /// @param name 名前
         /// @return その名前に対応するCellType値
-        [[nodiscard]] static CellType AddName(::std::string_view name);
+        [[nodiscard]] static CellType AddName(::std::string_view name, Graphics::Color color);
+
+        /// @brief 指定した名前に対応するCellType値を取得
+        /// @param name 名前
+        /// @return nameに対応するCellType値
+        /// @return 見つからない場合はInvalid()
+        [[nodiscard]] static CellType FromName(::std::string_view name) noexcept;
         
         CellType() = default;
+
+        [[nodiscard]] constexpr Graphics::Color Color() const noexcept;
 
         /// @brief 
         [[nodiscard]] constexpr uint32_t Id() const noexcept;
@@ -68,9 +77,14 @@ namespace CellSim::Cells
         return CellType(0);
     }
 
-    inline CellType CellType::AddName(const char* name)
+    inline CellType CellType::AddName(const char* name, Graphics::Color color)
     {
-        return AddName(::std::string_view{ name });
+        return AddName(::std::string_view{ name }, color);
+    }
+
+    constexpr Graphics::Color CellType::Color() const noexcept
+    {
+        return s_names[m_id].second;
     }
 
     constexpr uint32_t CellType::Id() const noexcept
@@ -80,7 +94,7 @@ namespace CellSim::Cells
 
     constexpr ::std::string const& CellType::Name() const noexcept
     {
-        return s_names[m_id];
+        return s_names[m_id].first;
     }
 
     constexpr bool operator==(CellType left, CellType right) noexcept